refactor config to its own module
